Rename ufunc to _ufunc and make cunumeric.ufunc a type, not a module
