+++ cunumeric/_ufunc/__init__.py
@@ -12,3 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
+from .bit_twiddling import *
+from .comparison import *
+from .math import *
+from .floating import *
+from .trigonometric import *
+from .ufunc import ufunc
