+++ cunumeric/_ufunc/trigonometric.py
