+++ cunumeric/linalg/linalg.py
@@ -14,9 +14,9 @@
 #
 
 import numpy as np
+from cunumeric._ufunc.math import sqrt as _sqrt
 from cunumeric.array import convert_to_cunumeric_ndarray
 from cunumeric.module import ndarray
-from cunumeric.ufunc.math import sqrt as _sqrt
 
 
 def cholesky(a):
